Convert price values from float to int for database compatibility
- Convert all price values from validate_price() to int(float(price)) in creation_flow.py
- Updated 4 instances where custom_price or price_value are stored
- Changed type hint for selected_price parameter from Optional[float] to Optional[int] in invitation_manager.py
- Updated docstring to reflect price is now stored as integer

This ensures all prices are stored as integers to match the database schema.

+++ services/relationships/invitations/invitation_manager.py
@@ -219,14 +219,14 @@ class InvitationManager:
             return False, str(e)
 
     def send_client_fills_invitation(self, trainer_id: str, client_phone: str,
-                                     client_name: str, selected_price: Optional[float] = None) -> Tuple[bool, str]:
+                                     client_name: str, selected_price: Optional[int] = None) -> Tuple[bool, str]:
         """Send invitation to client who will fill their own profile
 
         Args:
             trainer_id: UUID of the trainer
             client_phone: Client's phone number
             client_name: Client's name
-            selected_price: Price per session (None if "discuss later")
+            selected_price: Price per session as integer (None if "discuss later")
 
         Returns:
             Tuple of (success: bool, message: str)
